Added initial backbone for this project
It can store, list, and remove items from the user's data file. It also prints out a help message if you enter `!help`. I put it in a Dockerfile becuase it's hard to get support for the older python discord client and this is the only way I've been able to guarantee it. Local testing works great, but I have not had the ability to test the Dockerfile yet becuase Docker doesn't run on WSL. Will test on my laptop